Modify Makefile to build rpms from local git clones.
The only modification to the spec files is to allow the Version and Release
to be passed in from Makefile.

The Makefile now generates dev snapshot RPMs of heat master. RPMs for actual
releases can be managed by release branches of heat-rpms.

I've made as few changes as possible to the heat-jeos RPMs as I'm about
to nuke the project, to be replaced by heat-cfntools